use float for PnP

+++ src/image/pair/imagepair.h
@@ -24,8 +24,8 @@ public:
   Mat_<double> essential;
 
   // Projection matrices
-  Mat_<double> projection_img1;
-  Mat_<double> projection_img2;
+  Mat_<float> projection_img1;
+  Mat_<float> projection_img2;
   // Rectified rotation matrices
   Mat_<double> rotation_rect_img1;
   Mat_<double> rotation_rect_img2;
